refactor(ui): Created a dedicated module for managing the ui.
BREAKING CHANGE: THe syntax for the configuration of mappings changed.
See the documentation `:help smuggler-configuration` for details.
